Fast hack to prepare for major bugfix that would have prevented subclassing Function like PyDOLFIN wants to do. The change involves adding an "UFL id" as a static variable in each class, which is done easily in classes.py:
 for c in all_ufl_classes:
    c._uflid = c

The second part is replacing all "dict" and "defaultdict"
objects in algorithms where UFLObject subclasses are used
as keys by new classes UFLTypeDict and UFLTypeDefaultDict
found in common.py, which uses "key._uflid" instead of "key"
for dict lookup. So when looking for "dolfin.Function" in
this kind of dict, in fact "dolfin.Function._uflid" is
used which translates to "ufl.Function._uflid" which
is really just "ufl.Function".

This additional class introduces some overhead,
which can be optimized away later.