Bugfix.
There is some inconsistency in the handling of products
of tensors, need to work it out before continuing on AD.
Most likely I will make Product only capable of representing
products of scalars, which reduces special cases.
Different kinds of products of tensors will be represented
by conversion to scalars through indexing, application
of products on scalars, and conversion to tensors
using ComponentTensor.