Slight change in behaviour in is_spatially_constant.
Now UFL will crash whenever some place in the
code is asking whether an expression with an
undefined cell is spatially constant, as that
is in fact undefined. This will give a hopefully
more informative error message at a slightly
earlier stage than before.

Expressions with None for cell are still
assumed spatially constant.

Coefficients with elements with degree 0 are
also assumed spatially constant.