Add some cell geometry quantities.
The idea is that we can make these accessible
throughout the symbolic pipeline. This has the
advantage that they can be included in forms,
used for computing mesh properties, avoid some
special casing in form compilers for these
geometric quantities by making them similar to
other geometric quantities, and make them more
accessible for testing of form compilers.

The quantities added are:
- LocalCoordinate (cell.xi)
- GeometryJacobi (cell.J)
- InverseGeometryJacobi (cell.Jinv)
- GeometryJacobiDeterminant (cell.detJ)

Note that these are still not handled by algorithms
or form compilers, and may be removed later if
this was a bad idea.