Work on conditionals: Now allowing use of <,>,<=,>=,!= in conditional expressions.
Evaluating < etc. as bool will result in an error to protect from misuse.

Evaluating !=/NE as bool will result in
'is this expression NOT equivalently represented',
the opposite of the '==' operation.

Evaluating EQ as bool will result in 
'is this expression equivalently represented',
but the '==' operator is not mapped to EQ (yet?)
for some significant technical issues.

While there are still some pitfalls here,
I think this is a step up from the previous behaviour.