Add new literal type AbstractSymbol.
This type is not exported to the user but can be imported from ufl.classes.
The intention is a type that the form compilers can use internally to
represent some quantity that ufl doesn't need to know the properties of.
E.g. this will only be used in the later stages of compilation and
ufl algorithms will not handle it because they don't know what it means.