Indexed extract component from ListTensor via FixedIndex
Add a simplification rule when building an Indexed: if the argument is
a ListTensor and the index is a FixedIndex, extract the component from
the ListTensor pointed to by the index.

Since this component might be an Indexed itself we need to be careful
to avoid infinite loops when Python calls the constructor of the newly
created object.