Implement change_to_local_grad algorithm.
Intended to be applied after preprocessing of forms, at least after expand_derivatives.

Works for multiple layered grads (i.e. grad(grad(grad(f)))) if
the domain is affine, otherwise only a single grad.

When someone wants to do grad(grad(f)) with f in a high order domain
(or quads actually) they'll have to speak up and maybe give a hand
in the implementation of that. It's not much, just a bit messy.