Add cargo property to Mesh.
This is the same as the old Domain.data().

The plan was to avoid this and make dolfin.Mesh inherit from ufl.Mesh,
but for now that plan has failed.

The new name 'cargo' is somewhat better than 'data' because 'mesh data'
is a concept in dolfin meaning something completely different.