Require restriction for reference normals
Although on C0 non-manifold meshes the physical normal is opposite
depending on which side of the facet we view it from, this is not the
case for reference normals.  To see this, consider the two-element mesh:

    3---0
    |\ -|
    | \ |
    |+ \|
    2---1

When transformed to reference space the cells are:

    3
    |\
    | \
    |+ \
    1---2

and

    3
    |\
    | \
    |- \
    0---1

Clearly the reference normal for the 1-3 facet is not flipped when
switching from the +-ve to the --ve side of the restriction.