EnrichedElement: Remove +/* warning
In the distant past, A + B => MixedElement([A, B]).  The change that
A + B => EnrichedElement([A, B]) was made in d622c74 (22 March 2010).
A warning was introduced in fcbc5ff (26 March 2010) that the meaning of
"+" had changed, and that users wanting a MixedElement should use "*"
instead.  People have, presumably, been seeing this warning for 6 1/2
years by now, so it's probably safe to remove.