Add call operator syntax to Form to replace arguments and coefficients.
This makes it easier to e.g. express the norm defined by
a bilinear form as a functional. Example usage:

  M = a(f, f)  # Equivalent to replace(a, {u: f, v: f})
  c = a(coefficients={f:1})  # Equivalent to replace(a, {f:1})

Add @ operator to Form: form @ f == action(form, f) (python 3.5+ only)

Drop warning in Form str and repr added previously to check
for usage in ffc and dolfin (found a couple and fixed them).
These are useful for debugging but can be expensive and shouldn't
be used unless necessary.