Fix simplification of Indexed in apply_derivatives
When we observe an expression of type as_tensor(C[kk], jj)[ii], the
derivative rulesets try to simplify this to C[ii]. We must be careful
here that we replace all matching indices. For example, if we build
the first component of the diagonal of a Jacobian:

    as_tensor(Jacobian(mesh)[i, i], (i,))[0]

then we should rewrite this to

   Jacobian(mesh)[0, 0]

The previous code only looked for the first occurrence of any ii in
kk, resulting instead in

   Jacobian(mesh)[0, i]

To fix this build a full replacement dict and reconstruct the entire
list of indices.