Fix algorithmic scaling with number of subdomains for __rmul__
Rather than delegating to Form.__add__ for constructing forms with
multiple subdomain ids, build the integrals once individually and then
construct a single Form at the end. This avoids an O(N^2 log N)
algorithm (due to a sorting of the integrals each time in
Form.__init__).

Before:

    import time
    for n in range(0, 1001, 100):
        ids = tuple(range(n))
        start = time.time()
        expr = f*dx(ids)
        end = time.time()
        print(f"{n} {end-start:.3f}s")

0 0.000s
100 0.021s
200 0.076s
300 0.156s
400 0.258s
500 0.396s
600 0.569s
700 0.759s
800 1.041s
900 1.290s
1000 1.761s

After:

0 0.000s
100 0.005s
200 0.007s
300 0.011s
400 0.013s
500 0.015s
600 0.022s
700 0.024s
800 0.026s
900 0.028s
1000 0.034s

Other parts of the pipeline are still slow with many subdomain ids,
but this is a start.