pullback: Refactor and generalise
For TensorElements of piola-mapped things, we apply the pullback
tensor-shape component-wise. This can be done with free indices rather
than one component at a time, which results in slightly nicer
expressions.

Additionally, handle symmetric tensor elements where the sub-element
has a non-identity mapping. In this case, we just need to apply the
appropriate mapping to each unique component in the symmetric tensor
and glue them together again.