Performance improvements for transformations (#69)
* Eagerly DAGify when comparing for equality

If we determine that two expressions are equal, we can DAGify by
replacing the operands of one with the operands of the other. This
significantly speeds up traversals for complex forms that have many
equal sub-terms.

* map_dag: Allow caller to provide caches

This will help reuse of transformed traversals when the same mapping
function is applied multiple times. This occurs, for example, in the
dispatching for derivatives in apply_derivatives.

* restrictions: Reuse caches between calls to restriction propagation

* derivatives: Reuse cache between calls to dispatched rulesets