rename namespace from ufl to ufl_legacy
ufl 2023.1 diverged sufficiently that it is no longer compatible with
old dolfin, and not readily patched to restore compatibility.

The issue is raised at https://github.com/FEniCS/ufl/issues/151 .
Discussion at #151 concluded that since ufl is used elsewhere
(e.g. firedrake), it is preferable to rename old ufl (2022.2.0) as
ufl_legacy, rather than to rename latest ufl as uflx (which was done
for ffc->ffcx)

This patch commences the change of namespace from ufl to ufl_legacy,
intended for merging into a new ufl_legacy branch based on the
2022.2.0 tag.